+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-08-12
+
+### Added
+
+- Add `status` argument to list all projects in the manifest and indicate if they are modified.
+- Add `list` argument to list all projects in the manifest.
+
 ## [0.3.1] - 2024-05-03
 
 ### Fixed